Some other fixes in ROUTE Add interim file in VLM grib download (not tested)
git-svn-id: svn://svn.code.sf.net/p/qtvlm/code/trunk@927 2387b160-4621-4e8a-be36-40ea10d72da2

+++ src/MainWindow.cpp
@@ -1823,7 +1823,7 @@ void MainWindow::setPilototoFromRoute(ROUTE *route)
         POI * poi=route->getPoiList().at(n);
         if(poi->getTimeStamp()==-1)
             break;
-        if(poi->getTimeStamp()+30<=QDateTime().currentDateTimeUtc().toTime_t()) continue;
+        if(poi->getTimeStamp()+30<=(int)QDateTime().currentDateTimeUtc().toTime_t()) continue;
         pois.append(poi);
     }
     route->setPilototo(false);
